Extract registry cleanup from BakeryConfig
Move clean_caches() and clean_temporary() out of config/config.py into
posit_bakery/registry_management/clean.py. Both functions now accept a
list of ImageTarget objects instead of reading from self.targets.

config/config.py drops its last imports of ghcr, timedelta,
ImageBuildStrategy, python_on_whales, and DockerException. It no
longer depends on registry_management or build internals.

- cli/clean.py calls the new standalone functions
- Test mocks updated to patch the new function locations
- _merge_sequential_build_metadata_files test removed (logic was
  inlined into build_targets in previous commit)
- Build retry test updated to import from image.build
